+++ imagedl/modules/sources/__init__.py
@@ -1,6 +1,7 @@
 '''initialize'''
 from .base import BaseImageClient
 from .bing import BingImageClient
+from .i360 import I360ImageClient
 from .baidu import BaiduImageClient
 from .google import GoogleImageClient
 from ..utils import BaseModuleBuilder
@@ -9,7 +10,8 @@ from ..utils import BaseModuleBuilder
 '''ImageClientBuilder'''
 class ImageClientBuilder(BaseModuleBuilder):
     REGISTERED_MODULES = {
-        'BingImageClient': BingImageClient, 'BaiduImageClient': BaiduImageClient, 'GoogleImageClient': GoogleImageClient
+        'BingImageClient': BingImageClient, 'BaiduImageClient': BaiduImageClient, 'GoogleImageClient': GoogleImageClient, 
+        'I360ImageClient': I360ImageClient,
     }