+++ tests/test_client.py
@@ -1,0 +1,12 @@
+'''
+Function:
+    Implementation of Uniform ImageClient
+Author:
+    Zhenchao Jin
+WeChat Official Account (微信公众号):
+    Charles的皮卡丘
+'''
+from imagedl import imagedl
+
+image_client = imagedl.ImageClient(image_source='BaiduImageClient')
+image_client.startcmdui()