+++ tests/test_image_clients.py
@@ -1,25 +1,30 @@
 '''
 Function:
-    Implementation of ImageClients test
+    Implementation of Dedicated ImageClient
 Author:
     Zhenchao Jin
 WeChat Official Account (微信公众号):
     Charles的皮卡丘
 '''
 from imagedl.modules.sources.bing import BingImageClient
+from imagedl.modules.sources.i360 import I360ImageClient
 from imagedl.modules.sources.baidu import BaiduImageClient
 from imagedl.modules.sources.google import GoogleImageClient
 
 
 # bing tests
 client = BingImageClient()
-image_infos = client.search('美女', search_limits=10, num_threadings=1)
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
+client.download(image_infos, num_threadings=1)
+# 360 tests
+client = I360ImageClient()
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)
 # baidu tests
 client = BaiduImageClient()
-image_infos = client.search('美女', search_limits=10, num_threadings=1)
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)
 # google tests
 client = GoogleImageClient()
-image_infos = client.search('美女', search_limits=10, num_threadings=1)
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)